requirements-rag.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
tqdm>=4.66.1
